improved backend validations, empty id, uuid check

+++ server/src/index.js
@@ -10,7 +10,8 @@ const uuid = require('uuid/v4');
 const config = require('./config');
 
 // Extras
-const { validateParams, buildNewTaskQuery, buildUpdateTaskQuery, findPresentKeys } = require('./js/tasks');
+const { validateParams, buildNewTaskQuery, buildUpdateTaskQuery, 
+        findPresentKeys, isUUIDv4 } = require('./js/tasks');
 
 // Initialization
 const app = express();
@@ -39,7 +40,8 @@ pgClient
       id UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
       title TEXT NOT NULL,
       details TEXT,
-      completed BOOLEAN DEFAULT FALSE
+      completed BOOLEAN DEFAULT FALSE,
+      created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
     );
     `
   )
@@ -50,9 +52,12 @@ pgClient
 // Get all to do list tasks
 app.get('/api/v1/tasks', async (req, res) => {
   pgClient
-    .query("SELECT * FROM tasks;")
+    .query("SELECT * FROM tasks ORDER BY completed ASC, created_at ASC;")
     .then((result) => {
-      res.status(200).json(result.rows);
+      res.json({
+        result: "success",
+        tasks: result.rows
+      });
     })
     .catch((err) => console.log(err)
   );
@@ -60,10 +65,23 @@ app.get('/api/v1/tasks', async (req, res) => {
 
 // Get a single todo task
 app.get('/api/v1/tasks/:id', async (req, res) => {
+
+  if( !isUUIDv4(req.params.id) ){
+    res.status(400).json({
+      result: 'failed',
+      error: "Not a valid uuid"
+    });
+    return;
+  }
+
   pgClient
     .query(`SELECT * FROM tasks WHERE id='${req.params.id}'::uuid;`)
     .then((result) => {
-      res.status(200).json(result.rows[0]);
+      let responseResult = result.rows.length === 0 ? "failed" : "success"
+      res.status(200).json({
+        result: responseResult,
+        task: result.rows[0]
+      });
     })
     .catch((err) => console.log(err)
   );
@@ -74,38 +92,52 @@ app.post('/api/v1/tasks', async (req, res) => {
   
   let { errors, isValid } = validateParams(req.body);
   if(!isValid) {
-    res.status(400).json(errors);
+    res.status(400).json({errors: errors});
     return;  
   }
   
   let {isTitlePresent} = findPresentKeys(req.body);
   if( !isTitlePresent ) {
     errors.push("No title found in body parameters");
-    res.status(400).json(errors);
+    res.status(400).json({errors: errors});
     return;
   }
 
   pgClient
     .query( buildNewTaskQuery(req.body) )
     .then((result) => {
-      res.status(200).send("success fully created task");
+      let responseResult = result.rowCount === 0 ?  "failed" : "success"
+      res.json({
+        result: responseResult,
+      });
     })
     .catch((err) => console.log(err));
 });
 
 // Update a todo task
 app.put('/api/v1/tasks/:id', async (req, res) => {
+
+  if( !isUUIDv4(req.params.id) ){
+    res.status(400).json({
+      result: 'failed',
+      error: "Not a valid uuid"
+    });
+    return;
+  }
   
   let { errors, isValid } = validateParams(req.body);
   if(!isValid) {
-    res.status(400).json(errors);
+    res.status(400).json({errors: errors});
     return;  
   }
 
   pgClient
     .query( buildUpdateTaskQuery(req.body, req.params.id) )
     .then((result) => {
-      res.status(200).send("Updated task.");
+      let responseResult = result.rowCount === 0 ?  "failed" : "success"
+      res.json({
+        result: responseResult,
+      });
     })
     .catch((err) => console.log(err)
   );
@@ -113,10 +145,22 @@ app.put('/api/v1/tasks/:id', async (req, res) => {
 
 // Delete a todo task route
 app.delete('/api/v1/tasks/:id', async (req, res) => {
+
+  if( !isUUIDv4(req.params.id) ){
+    res.status(400).json({
+      result: 'failed',
+      error: "Not a valid uuid"
+    });
+    return;
+  }
+  
   pgClient
     .query(`DELETE FROM tasks WHERE id='${req.params.id}'::uuid;`)
     .then((result) => {
-      res.status(200).send("Successfully deleted task.");
+      let responseResult = result.rowCount === 0 ?  "failed" : "success"
+      res.json({
+        result: responseResult,
+      });
     })
     .catch((err) => console.log(err)
   );
